requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pdfplumber==0.10.3
google-generativeai==0.3.2
celery==5.3.6
redis==5.0.1
beautifulsoup4==4.12.3
lxml==5.1.0 